Break some functions out
